Add waste metric for coin selection
Waste is a metric introduced by the BnB algorithm as part of its bounding
procedure. It can also be used to compare and select the most optimal
coin selection algorithm.

This commit introduces the waste metic as a tool to compare the different
solutions provided by the coin selection algorithms.

Co-authored-by: csralvall <46329881+csralvall@users.noreply.github.com>
Co-authored-by: afilini <3058409+afilini@users.noreply.github.com>